[ 48] Add Delete Mode
This commit adds a new set of tasks to the Automated Ingest Celery
application.

It also affords two new Events - data_obj_delete and coll_delete.
This comes with its own event handlers. The Delete Mode can be set
with a new event handler method named delete_mode which returns a
member of a new enum in utils named DeleteMode.